Replace `object.create` with shallow clone
Benchmarks suggest shallow cloning is faster or equal to `Object.create` for small objects, and
parse contexts are likely to be small (0-10 keys).

There is an argument to use `Object.create`, whose performance doesn't depend on the size of the
object.  This can be decided if/when there's enough use-cases to justify a decision.